Make the GenericSO mroe generic by allowing it to define a class and accessor for setting the value. Add an Agent plugin. Refactor to bring like files together.
