+++ Assets/Digital Painting/ScriptableObjects/Validation/Agent.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c205eba1bc5c1ea4e9ca53fc052778e0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
