+++ Assets/Digital Painting/ScriptableObjects/Validation/Agent/NormalSpeedSettingSO.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9bcc61d2cc172664687bf58d9f3671f7
+guid: 3cc3a260c8d31d747a6ba485076f3b48
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
