--- Assets/Digital Painting/Scripts/Environment/DarkSkyApi.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 63094e24e454df74b87c7b0247a17a7e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
