--- Assets/Digital Painting/Scripts/Environment/DummyWeatherSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1a57eb467788ced40b82ace5861e7d60
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
