+++ Assets/Digital Painting/Scripts/Plugin/Abstract.meta
