+++ Assets/Digital Painting/Scripts/Plugin/Abstract/AbstractDayNightCycle.cs.meta
