+++ Assets/Digital Painting/Scripts/Plugin/Abstract/AbstractDayNightProfile.cs.meta
