+++ Assets/Digital Painting/Scripts/Plugin/Abstract/AbstractPluginDefinition.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
+using wizardscode.digitalpainting;
 
 namespace wizardscode.plugin
 {
@@ -10,8 +11,10 @@ namespace wizardscode.plugin
     /// </summary>
     public abstract class AbstractPluginDefinition
     {
+        public enum PluginCategory { Agent, DayNightCycle, Weather, Other }
+
         /// <summary>
-        /// Get the name of type of the manager, that is the Type of the MonoBehaviour that
+        /// Get the Type of the plugin manager, that is the Type of the MonoBehaviour that
         /// is used by the `DigitalPaintingManager` to manage the plugin.
         /// </summary>
         /// <returns></returns>
@@ -26,10 +29,18 @@ namespace wizardscode.plugin
         public abstract String GetProfileTypeName();
 
         /// <summary>
-        /// The name of a class that we know exists in the plugin implementation.
+        /// The name of a class that we know exists in the plugin implementation. This is used to verify that
+        /// any dependent assets are installed. If the plugin is self-contained, that is it does not required
+        /// another asset to function correctly then this can be set to null.
         /// </summary>
         /// <returns></returns>
         public abstract string GetPluginImplementationClassName();
+
+        /// <summary>
+        /// Get a human readable name for this type of plugin use in the UI.
+        /// </summary>
+        /// <returns></returns>
+        public abstract PluginCategory GetCategory();
 
         /// <summary>
         /// Get a human readable name for use in the UI.
@@ -38,16 +49,18 @@ namespace wizardscode.plugin
         public abstract string GetReadableName();
 
         /// <summary>
-        /// Get a URL from which this asset can be retrieved. Normally this would be an asset store
-        /// page, but it could be somewhere else, such as a GitHub repo.
+        /// Get a URL from which the dependent asset can be retrieved. Normally this would be an asset store
+        /// page, but it could be somewhere else, such as a GitHub repo. If the plugin is self-contained and
+        /// does not depend on another asset this can be set to null.
         /// </summary>
-        /// <returns></returns>
+        /// <returns>Either the URL for retrieving an asset this plugin depends upon or null if there
+        /// is no such dependency.</returns>
         public abstract string GetURL();
 
         /// <summary>
         /// Tests to see if the plugin asset is present and ready for use. The default test
-        /// is to see if the class named in `PluginImplmentationClassName` is available in the
-        /// assembly.
+        /// is to see if the class named by `GetPluginImplmentationClassName` is available in the
+        /// assembly. If no class is named then true will be returned.
         /// </summary>
         /// <returns>True if the plugin asset is present and can be used, otherwise false.</returns>
         public virtual bool AvailableForUse
@@ -55,6 +68,11 @@ namespace wizardscode.plugin
             get
             {
                 string className = GetPluginImplementationClassName();
+                if(className == null)
+                {
+                    // there is no dependency on external assets, so it's avaialble
+                    return true;
+                }
                 IEnumerable<Type> types = (from assembly in AppDomain.CurrentDomain.GetAssemblies()
                                            from type in assembly.GetTypes()
                                            where type.Name == className
@@ -72,6 +90,28 @@ namespace wizardscode.plugin
                     return false;
                 }
             }
+        }
+
+        private DigitalPaintingManager _dpManager;
+        private DigitalPaintingManager DigitalPaintingManager
+        {
+            get
+            {
+                if (_dpManager == null)
+                {
+                    _dpManager = GameObject.FindObjectOfType<DigitalPaintingManager>();
+                }
+                return _dpManager;
+            }
+        }
+
+
+        /// <summary>
+        /// Enable the plugin in your scene.
+        /// </summary>
+        public virtual void Enable()
+        {
+            DigitalPaintingManager.gameObject.AddComponent(GetManagerType());
         }
     }
 }
