+++ Assets/Digital Painting/Scripts/Plugin/Abstract/AbstractPluginDefinition.cs.meta
