+++ Assets/Digital Painting/Scripts/Plugin/Abstract/AbstractPluginProfile.cs
