+++ Assets/Digital Painting/Scripts/Plugin/Agent.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7c12afbddcf3bb14da2c1802bbf238be
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
