+++ Assets/Digital Painting/Scripts/Plugin/Agent/Agent_PluginManager.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace wizardscode.plugin
+{
+    public class Agent_PluginManager : AbstractPluginManager
+
+    {
+    }
+}
