+++ Assets/Digital Painting/Scripts/Plugin/Agent/Agent_PluginManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cff5af3938977f54080294ac1d62dfcd
+guid: be84058266bca5c47b798e1a9390a93d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
