+++ Assets/Digital Painting/Scripts/Plugin/DayNightCycle.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6518fbd90cf7e9849b80960ca698a990
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
