+++ Assets/Digital Painting/Scripts/Plugin/DayNightCycle/AbstractDayNightPluginDefinition.cs.meta
