+++ Assets/Digital Painting/Scripts/Plugin/DayNightCycle/AbstractPluginManager.cs.meta
