+++ Assets/Digital Painting/Scripts/Plugin/DayNightCycle/DayNightPluginManager.cs.meta
