+++ Assets/Digital Painting/Scripts/Plugin/DayNightCycle/SimpleDayNightPluginDefinition.cs.meta
