+++ Assets/Digital Painting/Scripts/Plugin/DayNightCycle/SimpleDayNightProfile.cs.meta
