+++ Assets/Digital Painting/Scripts/Plugin/DayNightCycle/ValidateSimpleDayNightProfile.cs
