+++ Assets/Digital Painting/Scripts/Plugin/Weather.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f633086bb5961c46a2f295148dac1bc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
