+++ Assets/Digital Painting/Scripts/Plugin/Weather/AbstractWeatherPluginDefinition.cs.meta
