+++ Assets/Digital Painting/Scripts/Plugin/Weather/AbstractWeatherProfile.cs
