+++ Assets/Digital Painting/Scripts/Plugin/Weather/DarkSkyApi.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 59a131da1a8d6634c81bbcd4ec6ab7bb
+guid: 63094e24e454df74b87c7b0247a17a7e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
