+++ Assets/Digital Painting/Scripts/Plugin/Weather/WeatherPluginManager.cs.meta
